requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
rich>=13.0.0
pyyaml>=6.0.1
scipy>=1.11.0
